Migrate check-sources.py for fontbakery 0.12
