+++ qa/check-sources.py
@@ -12,189 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import annotations
+from pathlib import Path
 
-from typing import Iterable, Tuple, Union
-
-from fontbakery.callable import check, condition
-from fontbakery.fonts_profile import profile_factory
-from fontbakery.message import Message
-from fontbakery.section import Section
-from fontbakery.status import FAIL, INFO, PASS, SKIP, WARN, Status
-from fontbakery import utils
-from fontTools.designspaceLib import DesignSpaceDocument
-from ufoLib2 import Font
-
-CheckStatus = Iterable[Tuple[Status, Union[Message, str]]]
-
-profile_imports = ()
-profile = profile_factory(default_section=Section("Google Sans Flex Source Checks"))
-
-
-@condition
-def ufo_font(ufo: str) -> Font:
-    return Font.open(ufo)
-
-
-@condition
-def ds(designspace: str) -> DesignSpaceDocument:
-    d = DesignSpaceDocument.fromfile(designspace)
-    d.loadSourceFonts(Font.open)
-    return d
-
-
-@check(id="com.google.fonts/check/googlesansflex/sources/same_tabular_width")
-def check_same_tabular_widths(ufo_font: Font, config) -> CheckStatus:
-    """Confirms that tabular glyphs have the same width within the same master."""
-
-    warnings_by_layer = {}
-    for layer in ufo_font.layers:
-        tabulars = {name for name in layer.keys() if ".tf" in name}
-        if not tabulars:
-            continue
-
-        if "zero.tf" in tabulars:
-            width_glyph = "zero.tf"
-        else:
-            width_glyph = next(iter(tabulars))
-        width = layer[width_glyph].width
-        if width is None:
-            yield (
-                FAIL,
-                f"layer {layer.name}: {width_glyph} has no width, stopping check",
-            )
-            return
-
-        for name in tabulars:
-            glyph = layer[name]
-            if glyph.width != width:
-                warnings_by_layer.setdefault(layer.name, []).append(
-                    (name, glyph.width, width_glyph, width)
-                )
-
-    if not warnings_by_layer:
-        yield PASS, "Tabular glyphs, if they exist, have the same width"
-    else:
-        for layer_name, warnings in warnings_by_layer.items():
-            yield (
-                FAIL,
-                Message(
-                    "mismatching-tabular-widths",
-                    f"layer '{layer_name}':\n\n"
-                    f"""{utils.bullet_list(config, [f"{name} ({width}) has different width than {ref_name} ({ref_width})"
-                     for (name, width, ref_name, ref_width) in warnings])}""",
-                ),
-            )
-
-
-@check(id="com.google.fonts/check/googlesansflex/sources/suspicious_kerning_values")
-def check_suspicious_kerning_values(ufo_font: Font, config) -> CheckStatus:
-    """Check for small and large kerning values outside a range and other
-    things."""
-
-    # Accept kerning values in the range [10, 200] for 1000 upM fonts.
-    threshold_low = round(10 * ufo_font.info.unitsPerEm / 1000)
-    threshold_high = round(200 * ufo_font.info.unitsPerEm / 1000)
-    threshold = range(threshold_low, threshold_high + 1)
-
-    def describe_pair(first: str, second: str) -> str:
-        glyphs = []
-        if first in ufo_font.groups:
-            glyphs.append(ufo_font.groups[first][0])
-        else:
-            glyphs.append(first)
-        if second in ufo_font.groups:
-            glyphs.append(ufo_font.groups[second][0])
-        else:
-            glyphs.append(second)
-        return "".join(f"/{name}" for name in glyphs)
-
-    suspicious_kerning = []
-    for (first, second), value in ufo_font.kerning.items():
-        # NOTE: Disable this for now, kerning can be cleaned up later.
-        # if value == 0:
-        #     if first in ufo_font.groups and second in ufo_font.groups:
-        #         yield WARN, f"Group-to-group pairs like {(first, second)} (e.g. {describe_pair(first, second)}) don't need zero values"
-        if value == 0:
-            continue
-        if abs(value) not in threshold:
-            suspicious_kerning.append(
-                ((first, second), describe_pair(first, second), value)
-            )
-
-    if not suspicious_kerning:
-        yield PASS, "No suspicion raised"
-    else:
-        yield (
-            WARN,
-            Message(
-                "suspicious-kerning-values",
-                f"Kerning values outside the accepted range of [{threshold_low}, {threshold_high}]:\n\n"
-                f"""{utils.bullet_list(
-                    config,
-                    [
-                        f"Pair {pair} (e.g. {example}): {value}"
-                        for (pair, example, value) in suspicious_kerning
-                    ],
-                )}""",
-            ),
-        )
-
-
-@check(id="com.google.fonts/check/googlesansflex/sources/same_kerning_groups")
-def check_same_kerning_groups(ds: DesignSpaceDocument) -> CheckStatus:
-    """Confirms that all sources have the same kerning groups per Designspace."""
-
-    default_source = ds.findDefault()
-    reference = default_source.font.groups
-    for source in ds.sources:
-        if source is default_source:
-            continue
-        if source.font.groups == reference:
-            yield PASS, f"{source.filename} has same kerning groups as default source"
-        else:
-            yield (
-                WARN,
-                f"{source.filename} does not have the same kerning groups as default source",
-            )
-
-
-@check(id="com.google.fonts/check/googlesansflex/sources/kerning_present")
-def check_kerning_present(ds: DesignSpaceDocument) -> CheckStatus:
-    """Check how much kerning pairs a source has, not counting exceptions."""
-
-    for source in ds.sources:
-        if "Skateboard" in source.filename:
-            yield SKIP, f"Skipping {source.filename} because it is experimental"
-            continue
-        if source.layerName is not None:
-            yield SKIP, f"Skipping {source.filename} because it is a sparse layer"
-            continue
-
-        ufo_font: Font = source.font
-        effective_kerning = {}
-        for (first, second), value in ufo_font.kerning.items():
-            # Skip exceptions for now, as we're more interested in the
-            # non-exceptions that make them meaningful.
-            if value == 0:
-                continue
-            # Kerning of non-existent things, skip.
-            if (first not in ufo_font.groups and first not in ufo_font) or (
-                second not in ufo_font.groups and second not in ufo_font
-            ):
-                continue
-            effective_kerning[(first, second)] = value
-
-        if effective_kerning:
-            yield (
-                INFO,
-                f"{source.filename}: Found {len(effective_kerning)} kerning pairs (not counting exceptions)",
-            )
-        else:
-            yield (
-                WARN,
-                f"{source.filename}: Found no kerning pairs (not counting exceptions)",
-            )
-
-
-profile.auto_register(globals())
+PROFILE = {
+    "check_definitions": [Path(__file__).parent / "checks" / "sources.py"],
+    "sections": {
+        "Google Sans Flex Source Checks": [
+            "com.google.fonts/check/googlesansflex/sources/same_tabular_width",
+            "com.google.fonts/check/googlesansflex/sources/suspicious_kerning_values",
+            "com.google.fonts/check/googlesansflex/sources/same_kerning_groups",
+            "com.google.fonts/check/googlesansflex/sources/kerning_present",
+        ]
+    },
+}
